git-svn-id: http://scripts.sil.org/svn-public/xetex/TRUNK@986 9f0520aa-a7ea-0310-80d7-d2976a5bef54

+++ texk/web2c/xetexdir/XeTeXFontMgr.cpp
@@ -38,8 +38,8 @@ authorization from the copyright holders.
 
 #include "XeTeXswap.h"
 
-#include "Features.h"
-#include "GlyphPositioningTables.h"
+#include "layout/ICUFeatures.h"
+#include "layout/GlyphPositioningTables.h"
 
 #include "sfnt.h"
 
@@ -586,7 +586,7 @@ XeTeXFontMgr::getOpSizeRecAndStyleFlags(Font* theFont)
 
 		const POSTTable* postTable = (const POSTTable*)getFontTablePtr(font, LE_POST_TABLE_TAG);
 		if (postTable != NULL) {
-			theFont->slant = (int)(1000 * (tan(Fix2X(-SWAP(postTable->italicAngle)) * M_PI / 180.0)));
+			theFont->slant = (int)(1000 * (tan(Fix2X(-SWAP(UInt32(postTable->italicAngle))) * M_PI / 180.0)));
 		}
 		deleteFont(font);
 	}
